Refactor typography styles in Privacy Policy and Terms of Use pages; remove font-serif class for consistency
Remove unused Blog, Careers, and News pages; streamline project structure

Update Disclaimer component to remove unused imports and improve typography

Revise Footer and Navigation components for consistent typography and updated labels

Add AssociatesSection and PartnersSection components to showcase team members with detailed profiles

Create AboutSection, CTASection, HeroSection, PartnersSection, and ServicesSection for the home page layout

Update Tailwind CSS configuration to change font families for a more modern look
