+++ src/app/globals.css
@@ -1,8 +1,15 @@
-@import url('https://fonts.googleapis.com/css2?family=Open+Sans:ital,wght@0,300;0,400;0,500;0,600;0,700;0,800;1,300;1,400;1,500;1,600;1,700;1,800&display=swap');
-@import url('https://fonts.googleapis.com/css2?family=Playfair+Display:wght@400;700&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Inter:ital,opsz,wght@0,14..32,100..900;1,14..32,100..900&display=swap');
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
+
+body {
+  font-family: 'Inter', sans-serif;
+}
+
+h1, h2, h3, h4, h5, h6 {
+  font-family: 'Georgia', serif;
+}
 
 @layer base {
   :root {
@@ -56,7 +63,11 @@
 }
 
 .font-serif {
-  font-family: 'Playfair Display', serif;
+  font-family: 'Georgia', serif;
+}
+
+.font-sans {
+  font-family: 'Inter', sans-serif;
 }
 
 .animate-fade-in {
