--- src/app/{blog,careers,news}/page.tsx
@@ -1,12 +1,0 @@
-export default function Blog() {
-  return (
-    <div className="min-h-screen">
-      <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-12">
-        <h1 className="text-4xl font-bold text-gray-900 mb-8">Blog</h1>
-        <p className="text-xl text-gray-600 mb-8">
-          Coming soon - Legal insights, industry updates, and thought leadership from Nora Chambers.
-        </p>
-      </div>
-    </div>
-  )
-}
