+++ src/components/sections/home/CTASection.tsx
@@ -1,0 +1,41 @@
+'use client'
+
+import { motion } from 'framer-motion'
+import Link from 'next/link'
+import { Button } from '@/components/ui/button'
+import { ArrowRight } from 'lucide-react'
+
+export default function CTASection() {
+  return (
+    <section className="py-24 bg-gradient-to-r from-primary to-accent text-white">
+      <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 text-center">
+        <motion.div
+          initial={{ opacity: 0, y: 30 }}
+          whileInView={{ opacity: 1, y: 0 }}
+          viewport={{ once: true }}
+          transition={{ duration: 0.8 }}
+        >
+          <h2 className="text-4xl md:text-5xl font-bold mb-6">
+            Ready to Work With Us?
+          </h2>
+          <p className="text-xl mb-8 max-w-2xl mx-auto opacity-90">
+            Get in touch with our experienced legal team to discuss how we can help with your legal needs.
+          </p>
+          <div className="flex flex-col sm:flex-row gap-4 justify-center">
+            <Link href="/contact">
+              <Button size="lg" variant="outline" className="bg-white text-primary hover:bg-gray-100 border-white">
+                Schedule Consultation
+                <ArrowRight className="ml-2 h-4 w-4" />
+              </Button>
+            </Link>
+            <Link href="/about">
+              <Button size="lg" variant="ghost" className="text-white hover:bg-white/20">
+                View Our Work
+              </Button>
+            </Link>
+          </div>
+        </motion.div>
+      </div>
+    </section>
+  )
+}
